Rename constraints folder to constraint
