+++ wpilibc/src/main/native/include/frc/trajectory/constraint/TrajectoryConstraint.h
